Sprint 11

+++ app/src/main/java/ch/epfl/sweng/project/Database/DBHelper.java
@@ -18,7 +18,9 @@ import ch.epfl.sweng.project.Model.Track;
 
 /**
  * This class provides methods to interact with a SQLite local database.
- * It allows to store and retrieve Runs.
+ * It allows to store and retrieve Runs and Challenges.
+ *
+ * @author Pablo Pfister
  */
 public class DBHelper extends SQLiteOpenHelper {
     private SQLiteDatabase db;
@@ -37,7 +39,8 @@ public class DBHelper extends SQLiteOpenHelper {
 
     /**
      * The constructor of the class.
-     * @param context
+     *
+     * @param context to use to open or create the database
      */
     public DBHelper(Context context) {
         super(context, DATABASE_NAME, null, 1);
@@ -45,10 +48,6 @@ public class DBHelper extends SQLiteOpenHelper {
         db = getWritableDatabase();
     }
 
-    /**
-     * This method creates the tables.
-     * @param db
-     */
     @Override
     public void onCreate(SQLiteDatabase db) {
         String createEffortsTableQuery = "CREATE TABLE " + RUNS_TABLE_NAME + " ("
@@ -77,14 +76,12 @@ public class DBHelper extends SQLiteOpenHelper {
         db.execSQL(createChallengesTableQuery);
     }
 
-    /**
-     * This method updates the database cleaning the tables.
-     * @param db
-     * @param oldVersion number
-     * @param newVersion number
-     */
     @Override
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
+        if (db == null || oldVersion < 0 || newVersion < oldVersion) {
+            throw new IllegalArgumentException();
+        }
+
         String dropEffortsTableQuery = "DROP TABLE IF EXISTS " + RUNS_TABLE_NAME;
         db.execSQL(dropEffortsTableQuery);
 
@@ -98,20 +95,30 @@ public class DBHelper extends SQLiteOpenHelper {
     }
 
     /**
-     * Inserts an Run in the database.
-     * @param run
+     * Insert a Run in the database.
+     *
+     * @param run the run to insert
      * @return true if the insertion was successful, false otherwise
      */
     public boolean insert(Run run) {
+        if (run == null) {
+            throw new IllegalArgumentException();
+        }
+
         return insert(run, false) != -1;
     }
 
     /**
-     * Inserts an Challenge in the database.
-     * @param challenge
+     * Insert a Challenge in the database.
+     *
+     * @param challenge the challenge to insert
      * @return true if the insertion was successful, false otherwise
      */
     public boolean insert(Challenge challenge) {
+        if (challenge == null) {
+            throw new IllegalArgumentException();
+        }
+
         String opponentName = challenge.getOpponentName();
         Challenge.Type type = challenge.getType();
         double goal = challenge.getGoal();
@@ -130,20 +137,21 @@ public class DBHelper extends SQLiteOpenHelper {
             case ABORTED_BY_OTHER:
                 result = 3;
                 break;
+            default:
         }
 
         long myRunId = insert(challenge.getMyRun(), true);
         long opponentRunId = insert(challenge.getOpponentRun(), true);
 
-        ContentValues contentValues = new ContentValues();
-        contentValues.put(CHALLENGES_COLS[1], opponentName);
-        contentValues.put(CHALLENGES_COLS[2], type.toString());
-        contentValues.put(CHALLENGES_COLS[3], goal);
-        contentValues.put(CHALLENGES_COLS[4], result);
-        contentValues.put(CHALLENGES_COLS[5], myRunId);
-        contentValues.put(CHALLENGES_COLS[6], opponentRunId);
+        ContentValues valuesToInsert = new ContentValues();
+        valuesToInsert.put(CHALLENGES_COLS[1], opponentName);
+        valuesToInsert.put(CHALLENGES_COLS[2], type.toString());
+        valuesToInsert.put(CHALLENGES_COLS[3], goal);
+        valuesToInsert.put(CHALLENGES_COLS[4], result);
+        valuesToInsert.put(CHALLENGES_COLS[5], myRunId);
+        valuesToInsert.put(CHALLENGES_COLS[6], opponentRunId);
 
-        long challengeId = db.insert(CHALLENGES_TABLE_NAME, null, contentValues);
+        long challengeId = db.insert(CHALLENGES_TABLE_NAME, null, valuesToInsert);
         if (challengeId != -1) {
             challenge.setId(challengeId);
             return true;
@@ -153,13 +161,14 @@ public class DBHelper extends SQLiteOpenHelper {
     }
 
     /**
-     * Inserts an Run in the database marking it as a challenge if needed.
-     * @param run
-     * @param isChallenge denotes if it's a challenge
+     * Insert a Run in the database marking it as a challenge if needed.
+     *
+     * @param run the run to insert
+     * @param isChallenge true if it's part of a challenge, false otherwise
      * @return the id of the inserted run
      */
     private long insert(Run run, boolean isChallenge) {
-        //insert all checkpoints
+        //insert all checkpoints that form the track
         Track track = run.getTrack();
         List<CheckPoint> checkpoints = track.getCheckpoints();
         long checkpointsFromId = -1;
@@ -175,38 +184,44 @@ public class DBHelper extends SQLiteOpenHelper {
         }
 
         //insert Run
-        String name = run.getName();
-        ContentValues runContentValues = new ContentValues();
+        ContentValues valuesToInsert = new ContentValues();
         int challenge = isChallenge ? 1 : 0;
-        runContentValues.put(RUNS_COLS[1], challenge);
-        runContentValues.put(RUNS_COLS[2], name);
-        runContentValues.put(RUNS_COLS[3], run.getDuration());
-        runContentValues.put(RUNS_COLS[4], checkpointsFromId);
-        runContentValues.put(RUNS_COLS[5], checkpointsToId);
-        long insertedRun = db.insert(RUNS_TABLE_NAME, null, runContentValues);
+        valuesToInsert.put(RUNS_COLS[1], challenge);
+        valuesToInsert.put(RUNS_COLS[2], run.getName());
+        valuesToInsert.put(RUNS_COLS[3], run.getDuration());
+        valuesToInsert.put(RUNS_COLS[4], checkpointsFromId);
+        valuesToInsert.put(RUNS_COLS[5], checkpointsToId);
+
+        long insertedRun = db.insert(RUNS_TABLE_NAME, null, valuesToInsert);
         run.setId(insertedRun);
         return insertedRun;
     }
 
     /**
-     * Inserts a checkpoint in the checkpoints table
+     * Insert a checkpoint in the checkpoints table.
+     *
      * @param checkpoint the checkpoint to insert
      * @return the id of the inserted row
      */
     private long insert(CheckPoint checkpoint) {
-        ContentValues contentValues = new ContentValues();
-        contentValues.put(CHECKPOINTS_COLS[1], checkpoint.getLatitude());
-        contentValues.put(CHECKPOINTS_COLS[2], checkpoint.getLongitude());
+        ContentValues valuesToInsert = new ContentValues();
+        valuesToInsert.put(CHECKPOINTS_COLS[1], checkpoint.getLatitude());
+        valuesToInsert.put(CHECKPOINTS_COLS[2], checkpoint.getLongitude());
 
-        return db.insert(CHECKPOINTS_TABLE_NAME, null, contentValues);
+        return db.insert(CHECKPOINTS_TABLE_NAME, null, valuesToInsert);
     }
 
     /**
-     * Deletes a given run
+     * Deletes a given run.
+     *
      * @param run to delete
      * @return true if the deletion was successful
      */
     public boolean delete(Run run) {
+        if (run == null) {
+            throw new IllegalArgumentException();
+        }
+
         long id = run.getId();
         if  (id >= 0) {
             String selectRun = RUNS_COLS[0] + " = " + id;
@@ -226,11 +241,16 @@ public class DBHelper extends SQLiteOpenHelper {
     }
 
     /**
-     * Deletes a given challenge
+     * Deletes a given challenge.
+     *
      * @param challenge to delete
      * @return true if the deletion was successful
      */
     public boolean delete(Challenge challenge) {
+        if (challenge == null) {
+            throw new IllegalArgumentException();
+        }
+
         long id = challenge.getId();
         if  (id >= 0) {
             delete(challenge.getMyRun());
@@ -244,7 +264,8 @@ public class DBHelper extends SQLiteOpenHelper {
     }
 
     /**
-     * Retrieves all runs present in the database.
+     * Retrieve all runs present in the database.
+     *
      * @return the list of runs present in the database
      */
     public List<Run> fetchAllRuns() {
@@ -266,7 +287,8 @@ public class DBHelper extends SQLiteOpenHelper {
     }
 
     /**
-     * Retrieves all challenges present in the database.
+     * Retrieve all challenges present in the database.
+     *
      * @return the list of challenges present in the database
      */
     public List<Challenge> fetchAllChallenges() {
@@ -286,7 +308,6 @@ public class DBHelper extends SQLiteOpenHelper {
 
                 double goal = result.getDouble(3);
 
-                boolean isWon = result.getShort(4) == 1;
                 Challenge.Result resultOfTheRun = Challenge.Result.LOST;
                 switch (result.getShort(4)) {
                     case 0:
@@ -301,6 +322,7 @@ public class DBHelper extends SQLiteOpenHelper {
                     case 3:
                         resultOfTheRun = Challenge.Result.ABORTED_BY_OTHER;
                         break;
+                    default:
                 }
 
                 Run myRun = fetchRun(result.getLong(5));
@@ -315,41 +337,36 @@ public class DBHelper extends SQLiteOpenHelper {
     }
 
     /**
-     * Fetches a specific run from the database given its id
-     * @param id
+     * Retrieve a specific run from the database given its id.
+     *
+     * @param id the id of the run to retrieve
      * @return the run
      */
     private Run fetchRun(long id) {
-        if (id < 0) {
-            throw new IllegalArgumentException();
-        }
-
         String selectRun = RUNS_COLS[0] + " = " + id;
         Cursor result = db.query(RUNS_TABLE_NAME, RUNS_COLS, selectRun, null, null, null, null);
         if (result.getCount() == 1) {
-            while (result.moveToNext()) {
-                String name = result.getString(2);
-                long duration = Long.parseLong(result.getString(3));
-                long fromId = result.getLong(4);
-                long toId = result.getLong(5);
-                Track track = fetchTrack(fromId, toId);
-                Run run = new Run(name, id);
-                run.setTrack(track);
-                run.setDuration(duration);
+            result.moveToNext();
+            String name = result.getString(2);
+            long duration = Long.parseLong(result.getString(3));
+            long fromId = result.getLong(4);
+            long toId = result.getLong(5);
+            Track track = fetchTrack(fromId, toId);
+            Run run = new Run(name, id);
+            run.setTrack(track);
+            run.setDuration(duration);
 
-                return run;
-            }
+            return run;
         } else {
             throw new NoSuchElementException();
         }
-
-        return null;
     }
 
     /**
-     * Retrieves a track from the database
-     * @param fromId
-     * @param toId
+     * Retrieve a track from the database given its starting and ending index.
+     *
+     * @param fromId the id of the first checkpoint
+     * @param toId the id of the last checkpoint
      * @return the track
      */
     private Track fetchTrack(long fromId, long toId) {
@@ -378,4 +395,12 @@ public class DBHelper extends SQLiteOpenHelper {
         return mContext.getDatabasePath(DATABASE_NAME);
     }
 
+    /**
+     * Getter for the database
+     *
+     * @return the database
+     */
+    public SQLiteDatabase getDatabase() {
+        return db;
+    }
 }
